feature: Component, service, model files are created.
Profile services file was created to handle call to API.
Actions file was updated with new actions.
Profile effects file is created.
Selectors is expanded to support profiles list.
Random users profile componet was implemented in the project to hadle Users Profiles path, where angular material table was used to display the results.
On silgle line click on the users table navigates to the profile page with the data from the row that user clicked